control_server/requirements.txt
pypjlink2
pyserial
wakeonlan
icmplib
